+++ src/main/metrics/index.ts
@@ -1,0 +1,6 @@
+export * from './counter';
+export * from './gauge';
+export * from './histogram';
+export * from './registry';
+export * from './global';
+export * from './util';