+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- feat: add APIs for Prometheus metrics
+
 ## 3.1.1
 
 - fix: child loggers merge data with context (previously their data was sitting in additional `data` key)